Unify cross-target release consumers and SDK/C-header authorities
Bind host WASI SDK assets and in-place target-qualified tools independently of native LLVM; consolidate target data models and exact installed native/WASM consumer receipts. Preserve unproved matrix cells and unresolved donor ABI work.

+++ runtime/molt-cpython-abi/src/abi_types.rs
@@ -342,7 +342,7 @@ pub struct PyTypeObject {
     pub tp_subclasses: *mut c_void,
     pub tp_weaklist: *mut PyObject,
     pub tp_del: Option<unsafe extern "C" fn(*mut PyObject)>,
-    pub tp_version_tag: c_ulong,
+    pub tp_version_tag: c_uint,
     pub tp_finalize: Option<unsafe extern "C" fn(*mut PyObject)>,
     pub tp_vectorcall: *mut c_void,
     pub tp_watched: u8,
